Move tables related browse code to separate folder
