+++ frontend/src/metabase/browse/containers/TableBrowser/TableBrowser.jsx
@@ -9,8 +9,8 @@ import { getSetting } from "metabase/selectors/settings";
 import { SAVED_QUESTIONS_VIRTUAL_DB_ID } from "metabase-lib/v1/metadata/utils/saved-questions";
 import * as ML_Urls from "metabase-lib/v1/urls";
 
-import TableBrowser from "../../components/TableBrowser";
 import { RELOAD_INTERVAL } from "../../constants";
+import TableBrowser from "../../tables/TableBrowser";
 
 const getDatabaseId = (props, { includeVirtual } = {}) => {
   const { params } = props;
