+++ frontend/src/metabase/browse/index.tsx
@@ -1,5 +1,5 @@
 export { BrowseMetrics } from "./metrics";
 export { BrowseModels } from "./models";
 export { BrowseDatabases } from "./databases";
-export { BrowseTables } from "./components/BrowseTables";
+export { BrowseTables } from "./tables";
 export { BrowseSchemas } from "./schemas";
