+++ frontend/src/metabase/browse/tables/TableBrowser/TableBrowser.styled.tsx
