+++ frontend/src/metabase/browse/tables/TableBrowser/index.js
