+++ frontend/src/metabase/browse/tables/analytics.ts
