+++ frontend/src/metabase/browse/tables/index.tsx
@@ -1,0 +1,1 @@
+export { BrowseTables } from "./BrowseTables";
